+++ src/EventStore.Client.Common/IEventFilter.cs
@@ -1,0 +1,12 @@
+namespace EventStore.Client {
+#if EVENTSTORE_GRPC_PUBLIC
+	public
+#else
+	internal
+#endif
+		interface IEventFilter {
+		PrefixFilterExpression[] Prefixes { get; }
+		RegularFilterExpression Regex { get; }
+		uint? MaxSearchWindow { get; }
+	}
+}
